requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
Flask==2.2.2
fonttools==4.37.4
greenlet==1.1.3
humanize==4.4.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
mqtt==0.0.1
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
paho-mqtt==1.6.1
pandas==1.3.5
Pillow==9.2.0
plotly==5.10.0
portpicker==1.5.2
psutil==5.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
tenacity==8.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
wget==3.2
zipp==3.8.1
